- promise wrapper for identity creation callbacks added
